refactor: Issue #2159 - Refactor staging validation test suites to import production modules
## Summary
Completed refactoring of staging validation test suites to import production modules instead of using inline implementations:

### Changes Made

**1. Created staging-validation-helpers.js**
   - Extracted inline helper functions into a shared module
   - Functions: validateAudit, runPerformanceBench, testErrorScenarios, validateReports, validateIntegrity, parseArguments, executeAllValidations
   - Fixed NaN issue by adding guard against empty benchmarks
   - Uses nullish coalescing operator (??) for consistency/validity scores
   - Proper radix handling in parseArguments (radix: 10)

**2. Created integration-workflow-staging-helpers.js**
   - Extracted staging environment validation functions
   - Functions: validateStagingEnvironment, runPerformanceBench, validateDataIntegrity, runComplianceChecks, calculateReadinessScore, canPromoteToProduction, generateStagingReport
   - Consistent use of nullish coalescing operator for optional properties
   - Comprehensive production readiness workflow helpers

**3. Refactored staging-validation.test.js**
   - Removed 183 lines of inline implementations
   - Now imports all functions from staging-validation-helpers.js
   - All 60 tests passing ✅
   - Tests now test production module code instead of inline mocks

**4. Refactored integration-workflow-staging.test.js**
   - Removed 304 lines of inline implementations
   - Now imports all functions from integration-workflow-staging-helpers.js
   - All 28 tests passing ✅
   - Tests validate staging to production promotion workflow

**5. Updated staging-validation.js**
   - Added imports for helper functions from staging-validation-helpers.js
   - Production module now references shared helpers

## Test Results
- staging-validation.test.js: 60 tests passing ✅
- integration-workflow-staging.test.js: 28 tests passing ✅
- **Total: 88 tests passing ✅**

## Bugs Fixed
- Fixed NaN averages when benchmarks array is empty
- Proper nullish coalescing (??) for optional configuration values
- Correct radix specification in numeric parsing

## Impact
- Eliminated code duplication between tests and production
- Improved maintainability with shared helper modules
- Better alignment with production code patterns
- Reduced test suite size by ~500 lines through imports
